backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
passlib==1.7.4
bcrypt==4.2.1
python-jose[cryptography]
pydantic[email]
python-multipart